+++ src/ingest_validation_tools/validation_utils.py
@@ -14,7 +14,7 @@ from ingest_validation_tools.directory_validator import (
 from ingest_validation_tools.schema_loader import (
     PreflightError,
     SchemaVersion,
-    get_directory_schema,
+    get_possible_directory_schemas,
 )
 from ingest_validation_tools.table_validator import ReportType
 
@@ -158,7 +158,7 @@ def get_data_dir_errors(
     root_path: Path,
     data_dir_path: Path,
     dataset_ignore_globs: List[str] = [],
-) -> Optional[dict]:
+) -> Dict[str, Union[List[str], str]]:
     """
     Validate a single data_path.
     """
@@ -178,39 +178,55 @@ def get_data_dir_errors(
         data_paths = list(shared_directories)
     # Otherwise, do nothing we can just use the predefine data_path
 
-    schema = get_directory_schema(dir_schema=dir_schema)
+    possible_schemas = get_possible_directory_schemas(dir_schema)
 
-    if schema is None:
-        return {"Undefined directory schema": dir_schema}
+    if possible_schemas is None:
+        return {dir_schema: ["No matching directory schemas found."]}
 
-    schema_warning_fields = [field for field in schema if field in ["deprecated", "draft"]]
-    schema_warning = (
-        {f"{schema_warning_fields[0].title()} directory schema": dir_schema}
-        if schema_warning_fields
-        else None
+    # Collect errors, discard if schema validates against a minor version
+    errors = []
+
+    # Make sure possible_schemas is sorted by key (descending) to evaluate highest minor version first
+    for schema_name, schema in sorted(possible_schemas.items(), reverse=True):
+        schema_errors = defaultdict(list)
+        schema_warning_fields = [field for field in schema if field in ["deprecated", "draft"]]
+        schema_warning = (
+            f"{schema_warning_fields[0].title()} directory schema: {schema_name}"
+            if schema_warning_fields
+            else None
+        )
+
+        try:
+            validate_directory(
+                data_paths, schema["files"], dataset_ignore_globs=dataset_ignore_globs
+            )
+        except DirectoryValidationErrors as e:
+            # If there are DirectoryValidationErrors and the schema is deprecated/draft...
+            #    schema deprecation/draft status is more important.
+            if schema_warning:
+                schema_errors[schema_name].append(schema_warning)
+            else:
+                schema_errors[schema_name].append(e.errors)
+        except OSError as e:
+            # If there are OSErrors and the schema is deprecated/draft...
+            #    the OSErrors are more important.
+            schema_errors[schema_name].append(f"{e.strerror}: {e.filename}")
+        if schema_errors:
+            errors.append(schema_errors)
+            continue
+        elif schema_warning:
+            errors.append({schema_name: schema_warning})
+            continue
+        # Found a schema with no problems!
+        # Throw away any found errors.
+        return {schema_name: "No errors!"}
+    # Did not find a schema that validated;
+    # return first (highest) schema version errors.
+    if errors:
+        return errors[0]
+    raise Exception(
+        f"Unknown error validating directory schema: {[x.as_posix() for x in data_paths]}"
     )
-
-    printable_data_paths = [x.as_posix() for x in data_paths]
-
-    try:
-        validate_directory(data_paths, schema["files"], dataset_ignore_globs=dataset_ignore_globs)
-    except DirectoryValidationErrors as e:
-        # If there are DirectoryValidationErrors and the schema is deprecated/draft...
-        #    schema deprecation/draft status is more important.
-        if schema_warning:
-            return schema_warning
-        errors = {}
-        errors[f"{','.join(printable_data_paths)} (as {dir_schema})"] = e.errors
-        return errors
-    except OSError as e:
-        # If there are OSErrors and the schema is deprecated/draft...
-        #    the OSErrors are more important.
-        return {f"{','.join(printable_data_paths)} (as {dir_schema})": {e.strerror: e.filename}}
-    if schema_warning:
-        return schema_warning
-
-    # No problems!
-    return None
 
 
 def get_context_of_decode_error(e: UnicodeDecodeError) -> str: